Move validators to Web api project
